# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kattraan LMS API",
    "version": "2.0.0",
    "description": "Comprehensive API documentation for Kattraan LMS.\n\n**Authentication**:\nThis API uses **HttpOnly Cookies** for authentication (`accessToken` and `refreshToken`).\n\n**Security Features**:\n- UUID-based role system (roles are NOT exposed in responses)\n- Token rotation on refresh\n- Session management (max 3 concurrent sessions)\n- Token blacklisting on logout\n- Rate limiting on auth endpoints\n\n**Important**: To test protected endpoints, you must first call `/auth/login`. The browser will automatically attach cookies to subsequent requests."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local (port 5000)"
    },
    {
      "url": "http://localhost:5001/api",
      "description": "Local (port 5001, default)"
    }
  ],
...
  "paths": {
...
    "/media/delete/{key}": {
      "delete": {
        "tags": ["Media"],
        "summary": "Delete a file",
        "parameters": [
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "File deleted" }
        }
      }
    },
    "/media/bulk-upload": {
      "post": {
        "tags": ["Media"],
        "summary": "Bulk upload files",
        "description": "Upload up to 10 files. Requires instructor or admin.",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "type": "object", "properties": { "files": { "type": "array", "items": { "type": "string", "format": "binary" } } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Files uploaded" },
          "400": { "description": "No files provided" },
          "401": { "description": "Not authenticated" }
        }
      }
    },
...
    "/sections": {
      "get": { "tags": ["Sections"], "summary": "List sections (optional query: course)", "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }], "parameters": [ { "name": "course", "in": "query", "schema": { "type": "string" } } ], "responses": { "200": { "description": "List of sections" } } },
      "post": { "tags": ["Sections"], "summary": "Create section", "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }], "responses": { "201": { "description": "Section created" } } }
    },
...
    "/chapters": {
      "get": { "tags": ["Chapters"], "summary": "List chapters", "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }], "parameters": [ { "name": "section", "in": "query", "schema": { "type": "string" } } ], "responses": { "200": { "description": "List of chapters" } } },
      "post": { "tags": ["Chapters"], "summary": "Create chapter", "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }], "responses": { "201": { "description": "Chapter created" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (HS256). Use: Authorization: Bearer <token>"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly cookie containing JWT access token"
      }
    }
  }
}